pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "customer_segmentation"
version = "0.1.0"
description = "Customer segmentation project"
authors = [
  { name="Belinda Freiheit", email="belinda.freiheit@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
  "numpy",
  "pandas",
  "matplotlib",
  "seaborn",
  "scikit-learn",
  "scipy",
  "python-dotenv",
  "psycopg2-binary",
  "SQLAlchemy"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]